chore: rename `_container` folder to `container`
