+++ src/app/(auth)/forgot-password/otp-verification/layout.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Title from '../../_container/Title';
+import Title from '../../container/Title';
 
 export default function OTPVerificationLayout({
   children,
