+++ src/app/(auth)/login/loading.tsx
@@ -1,7 +1,7 @@
 import { Button } from 'antd';
 import Link from 'next/link';
 import { SkeletonInput } from '@/shared/container/ClientAntd/Skeleton/Skeleton';
-import OAuthLoginButtons from '../_container/OAuthLoginButtons';
+import OAuthLoginButtons from '../container/OAuthLoginButtons';
 
 export default function LoginLoading() {
   return (
