+++ src/app/(auth)/register/create-profile/layout.tsx
@@ -1,4 +1,4 @@
-import Title from '../../_container/Title';
+import Title from '../../container/Title';
 
 export default function CreateProfileLayout({
   children,
